chore: move identity entity in its own folder
