+++ src/api/private/tokens/tokens.controller.spec.ts
@@ -10,8 +10,8 @@ import { getRepositoryToken } from '@nestjs/typeorm';
 import { AuthToken } from '../../../auth/auth-token.entity';
 import { AuthModule } from '../../../auth/auth.module';
 import appConfigMock from '../../../config/mock/app.config.mock';
+import { Identity } from '../../../identity/identity.entity';
 import { LoggerModule } from '../../../logger/logger.module';
-import { Identity } from '../../../users/identity.entity';
 import { Session } from '../../../users/session.entity';
 import { User } from '../../../users/user.entity';
 import { TokensController } from './tokens.controller';
