+++ src/media/media.service.spec.ts
@@ -15,6 +15,7 @@ import { Author } from '../authors/author.entity';
 import mediaConfigMock from '../config/mock/media.config.mock';
 import { ClientError, NotInDBError } from '../errors/errors';
 import { Group } from '../groups/group.entity';
+import { Identity } from '../identity/identity.entity';
 import { LoggerModule } from '../logger/logger.module';
 import { Note } from '../notes/note.entity';
 import { NotesModule } from '../notes/notes.module';
@@ -23,7 +24,6 @@ import { NoteGroupPermission } from '../permissions/note-group-permission.entity
 import { NoteUserPermission } from '../permissions/note-user-permission.entity';
 import { Edit } from '../revisions/edit.entity';
 import { Revision } from '../revisions/revision.entity';
-import { Identity } from '../users/identity.entity';
 import { Session } from '../users/session.entity';
 import { User } from '../users/user.entity';
 import { UsersModule } from '../users/users.module';
