+++ src/users/users.module.ts
@@ -6,8 +6,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { Identity } from '../identity/identity.entity';
 import { LoggerModule } from '../logger/logger.module';
-import { Identity } from './identity.entity';
 import { Session } from './session.entity';
 import { User } from './user.entity';
 import { UsersService } from './users.service';
